+++ src/PhpWord/Writer/Word2007/Style/Font.php
@@ -90,6 +90,10 @@ class Font extends AbstractStyle
             $xmlWriter->writeAttributeIf($language->getLatin() !== null, 'w:val', $language->getLatin());
             $xmlWriter->writeAttributeIf($language->getEastAsia() !== null, 'w:eastAsia', $language->getEastAsia());
             $xmlWriter->writeAttributeIf($language->getBidirectional() !== null, 'w:bidi', $language->getBidirectional());
+            //if bidi is not set but we are writing RTL, write the latin language in the bidi tag
+            if ($style->isRTL() && $language->getBidirectional() === null && $language->getLatin() !== null) {
+                $xmlWriter->writeAttribute('w:bidi', $language->getLatin());
+            }
             $xmlWriter->endElement();
         }